+++ packages/expo-modules-core/common/cpp/fabric/ExpoViewShadowNode.cpp
@@ -33,10 +33,10 @@ void ExpoViewShadowNode::initialize() noexcept {
   }
 
   if (YGNodeStyleGetDisplay(&yogaNode_) == YGDisplayContents) {
-    auto it = viewProps.propsMap.find("forceFlattenDisplayContents");
-    bool forceFlattenDisplayContents = (it != viewProps.propsMap.end()) ? it->second.getBool() : true;
+    auto it = viewProps.propsMap.find("disableForceFlatten");
+    bool disableForceFlatten = (it != viewProps.propsMap.end()) ? it->second.getBool() : false;
     
-    if (!forceFlattenDisplayContents) {
+    if (disableForceFlatten) {
       traits_.unset(react::ShadowNodeTraits::Trait::ForceFlattenView);
     }
   }